server/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.3.7
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
iniconfig==1.1.1
node==0.9.28
odict==1.8.0
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
plumber==1.6
psycopg2-binary==2.9.1
py==1.11.0
pyjwt==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dotenv==0.19.2
pytz==2021.3
sh==1.12.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
watchman==0.0.8
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
